ENH: Check for whitespace
