+++ codespell_lib/tests/test_basic.py
@@ -6,7 +6,6 @@ import contextlib
 import os
 import os.path as op
 import subprocess
-import re
 import sys
 import tempfile
 import warnings
@@ -396,37 +395,6 @@ def FakeStdin(text):
         yield
     finally:
         sys.stdin = oldin
-
-
-def test_dictionary_formatting():
-    """Test that all dictionary entries are in lower case and non-empty."""
-    err_dict = dict()
-    with open(op.join(op.dirname(__file__), '..', 'data',
-                      'dictionary.txt'), 'rb') as fid:
-        for line in fid:
-            err, rep = line.decode('utf-8').split('->')
-            err = err.lower()
-            assert err not in err_dict, 'entry already exists'
-            rep = rep.rstrip('\n')
-            assert len(rep) > 0, ('%s: correction %r must be non-empty'
-                                  % (err, rep))
-            assert not re.match('^\s.*', rep), ('%s: correction %r cannot '
-                                                'start with whitespace'
-                                                % (err, rep))
-            if rep.count(','):
-                if not rep.endswith(','):
-                    assert 'disabled' in rep.split(',')[-1], \
-                        ('currently corrections must end with trailing "," (if'
-                         ' multiple corrections are available) or '
-                         'have "disabled" in the comment')
-            err_dict[err] = rep
-            reps = [r.strip() for r in rep.lower().split(',')]
-            reps = [r for r in reps if len(r)]
-            unique = list()
-            for r in reps:
-                if r not in unique:
-                    unique.append(r)
-            assert reps == unique, 'entries are not (lower-case) unique'
 
 
 def test_case_handling(reload_codespell_lib):
